separate nets out
