+++ docs/core.html
@@ -57,7 +57,7 @@
 <p><strong>rank</strong> (<em>int</em>) - Rank of DeepK operation. Use 0 for full rank. Will be set to <code>min(num_encoded_states, num_training_samples-1)</code> if the provided value is greater.</p>
 </li>
 <li>
-<p>Parameters required by <a href="https://galoisinc.github.io/deep-koopman/utils.html#deepk.utils.AutoEncoder">utils.AutoEncoder</a>:</p>
+<p>Parameters required by <a href="https://galoisinc.github.io/deep-koopman/nets.html#deepk.nets.AutoEncoder">AutoEncoder</a>:</p>
 <ul>
 <li>
 <p><strong>num_encoded_states</strong> (<em>int</em>).</p>
@@ -131,7 +131,7 @@
 <p><strong>tscale</strong> (<em>float</em>) - All t data is scaled by <code>tscale</code> so that <code>ttr</code> becomes <code>[0,1,&hellip;]</code></p>
 </li>
 <li>
-<p><strong>net</strong> (<em>utils.AutoEncoder</em>) - DeepKoopman neural network. <code>num_input_states</code> is set by num_input_states in X data, while other parameters are passed via this class.</p>
+<p><strong>net</strong> (<em>nets.AutoEncoder</em>) - DeepKoopman neural network. <code>num_input_states</code> is set by num_input_states in X data, while other parameters are passed via this class.</p>
 </li>
 <li>
 <p><strong>opt</strong> (<em>torch optimizer</em>) - Optimizer used to train DeepKoopman neural net.</p>
