+++ docs/index.html
@@ -47,6 +47,10 @@
 <dd>
 <div class="desc"><p>Loss functions used for DeepKoopman neural net optimization.</p></div>
 </dd>
+<dt><code class="name"><a title="deepk.nets" href="nets.html">deepk.nets</a></code></dt>
+<dd>
+<div class="desc"><p>Neural nets used inside DeepKoopman.</p></div>
+</dd>
 <dt><code class="name"><a title="deepk.utils" href="utils.html">deepk.utils</a></code></dt>
 <dd>
 <div class="desc"><p>Utilities</p></div>
@@ -73,6 +77,7 @@
 <li><code><a title="deepk.errors" href="errors.html">deepk.errors</a></code></li>
 <li><code><a title="deepk.hyp_search" href="hyp_search.html">deepk.hyp_search</a></code></li>
 <li><code><a title="deepk.losses" href="losses.html">deepk.losses</a></code></li>
+<li><code><a title="deepk.nets" href="nets.html">deepk.nets</a></code></li>
 <li><code><a title="deepk.utils" href="utils.html">deepk.utils</a></code></li>
 </ul>
 </li>
